+++ docs/api/changelog.rst
@@ -362,6 +362,8 @@ Version 0.10
 
   Relying on the field type to be present is thus a non-starter.
 
+* Update :meth:`aioxmpp.im.conversation.AbstractConversation.invite` interface.
+
 .. _api-changelog-0.9:
 
 Version 0.9